moved around the view components into their own folder, fixed small bug with rendering error message if not match
