+++ src/components/results/CommonData.js
@@ -1,5 +1,4 @@
 import React from "react"; 
-import PropTypes from "prop-types"; 
 import DataRow from "./DataRow"; 
 
 function FormatCommonData({commonData}) {
@@ -10,13 +9,5 @@ function FormatCommonData({commonData}) {
         </>
     ); 
 }
-
-// FormatCommonData.propTypes = {
-//     commonData: PropTypes.shape({
-//       _id: PropTypes.number,
-//       external_id: PropTypes.string,
-//       created_at: PropTypes.string,
-//     })
-// }; 
 
 export default FormatCommonData;
