+++ src/components/results/Result.css
