+++ src/components/results/Result.js
