+++ src/components/results/Tickets.js
